In the event list window, program change can be edited by a pop-up menu listing available patch names.
git-svn-id: svn+ssh://svn.osdn.jp/svnroot/alchemusica/trunk@112 19e2383c-df9d-4eb5-8775-ae43218f5bcb

+++ MD_package/MDEvent.h
@@ -350,7 +350,8 @@ enum {
 	kMDEventFieldInvalid
 };
 
-typedef long MDEventFieldDataWhole; /*  MDEventFieldData と少なくとも同じサイズの整数型  */
+/*  MDEventFieldData と少なくとも同じサイズの整数型  */
+typedef intptr_t MDEventFieldDataWhole;
 typedef union MDEventFieldData {
 	MDEventFieldDataWhole whole;			
 	long			longValue;		/*  code, data1  */
